+++ python/sglang/multimodal_gen/runtime/pipelines_core/schedule_batch.py
@@ -154,7 +154,7 @@ class Req:
 
     # results
     output: torch.Tensor | None = None
-    output_compression: float | None = None
+    output_compression: int | None = None
 
     def __init__(self, **kwargs):
         # Initialize dataclass fields
